+++ matrix-sdk/src/androidTest/java/org/matrix/androidsdk/lazyloading/RoomStateTest.java
@@ -70,7 +70,7 @@ public class RoomStateTest {
 
         Assert.assertEquals(1, aliceRoom.getNumberOfInvitedMembers());
         Assert.assertEquals(3, aliceRoom.getNumberOfJoinedMembers());
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     @Test
@@ -108,7 +108,7 @@ public class RoomStateTest {
 
         Assert.assertEquals(1, aliceRoom.getNumberOfInvitedMembers());
         Assert.assertEquals(3, aliceRoom.getNumberOfJoinedMembers());
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     @Test
@@ -167,7 +167,7 @@ public class RoomStateTest {
         });
         recursiveBackPaginate(liveTimeline, 0, 30, 120);
         mTestHelper.await(lock);
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     @Test
@@ -204,7 +204,7 @@ public class RoomStateTest {
 
         Assert.assertEquals(1, aliceRoom.getNumberOfInvitedMembers());
         Assert.assertEquals(3, aliceRoom.getNumberOfJoinedMembers());
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     @Test
@@ -270,7 +270,7 @@ public class RoomStateTest {
             }
         });
         mTestHelper.await(lock);
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     @Test
@@ -343,7 +343,7 @@ public class RoomStateTest {
             }
         });
         mTestHelper.await(lock);
-        mTestHelper.clearAllSessions(data);
+        mLazyLoadingTestHelper.clearAllSessions(data);
     }
 
     /**
